+++ HaselTweaks/Tweaks/EnhancedMiragePrismBoxConfig.cs
@@ -1,0 +1,15 @@
+namespace HaselTweaks.Tweaks;
+
+public class EnhancedMiragePrismBoxConfiguration
+{
+    public bool EnableAutoFillHandIn = false;
+}
+
+public partial class EnhancedMiragePrismBox
+{
+    public override void DrawConfig()
+    {
+        _configGui.DrawConfigurationHeader();
+        _configGui.DrawBool("EnableAutoFillHandIn", ref _config.EnableAutoFillHandIn);
+    }
+}