+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [44.0.0] (2026-05-13)
+
+### New Tweak: Enhanced Glamour Dresser
+
+Not so sure about this yet, but it's a start.
+
+Configuration options:
+
+- Automatically select missing items from the inventory when opening the Outfit Glamour Creation
+
+### Tweak reworked: Glamour Dresser Alert
+
+Previously called Glamour Dresser Armoire Alert.
+
+The window now also lists items that can be stored as outfit set.
+
 ## [43.0.0] (2026-05-11)
 
 ### New Tweak: Enhanced Crafting Log
@@ -2501,7 +2517,8 @@ Refreshes the material list and recipe tree when you've crafted or gathered an i
 
 Opens duty finder for the duty you clicked on in the Wondrous Tails Journal.
 
-[unreleased]: https://github.com/Haselnussbomber/HaselTweaks/compare/v43.0.0...main
+[unreleased]: https://github.com/Haselnussbomber/HaselTweaks/compare/v44.0.0...main
+[44.0.0]: https://github.com/Haselnussbomber/HaselTweaks/compare/v43.0.0...v44.0.0
 [43.0.0]: https://github.com/Haselnussbomber/HaselTweaks/compare/v42.1.3...v43.0.0
 [42.1.3]: https://github.com/Haselnussbomber/HaselTweaks/compare/v42.1.2...v42.1.3
 [42.1.2]: https://github.com/Haselnussbomber/HaselTweaks/compare/v42.1.1...v42.1.2